backend/requirements.txt
Django==5.1.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymysql==1.1.0
python-dotenv==1.0.0
